Removed update password feature, modified reset password feature
The code removed the update password feature as it was unwanted. Also, the reset password feature was modified to reset the password if the new password and the confirm password are equal and the old password provided is equal to the existing password. Finally, the endpoint route was also changed from "reset-password" to "changePassword".