Rearrange password validation checks in OwnSecurityServiceImpl
The order of password validation checks in the `ensureNewPasswordIsValid` method has been changed. Now, the method first checks if the new password and confirmation password are identical before checking if the new password is the same as the old one. This provides a more logical flow of the password validation process.